Adds documentation for Ruby Managers.
If Vim isn't well integrated with the Ruby Manager, it will probably use the system Ruby,
sometimes it can be an old version of Ruby causing errors (e.g. hash rocket
vs new syntax), but as the `:TagsGenerate` doesn't show errors to users
(`2>/dev/null`), this commit explains the steps to check if it is well
configured.